// App/Main-API/Client/src/app/common/enum/permission.enum.ts
export enum Permission {
    // Users
    CanViewUsers = 'CanViewUsers',
    BrowseUsers = 'BrowseUsers',
    EditUsers = 'EditUsers',
    DeleteUsers = 'DeleteUsers',
  
    // Shelters
    CanViewShelters = 'CanViewShelters',
    BrowseShelters = 'BrowseShelters',
    EditShelters = 'EditShelters',
    DeleteShelters = 'DeleteShelters',
  
    // Animals
    CanViewAnimals = 'CanViewAnimals',
    CreateAnimals = 'CreateAnimals',
    BrowseAnimals = 'BrowseAnimals',
    EditAnimals = 'EditAnimals',
    DeleteAnimals = 'DeleteAnimals',
  
    // Animal Types
    CanViewAnimalTypes = 'CanViewAnimalTypes',
    CreateAnimalTypes = 'CreateAnimalTypes',
    BrowseAnimalTypes = 'BrowseAnimalTypes',
    EditAnimalTypes = 'EditAnimalTypes',
    DeleteAnimalTypes = 'DeleteAnimalTypes',
  
    // Breeds
    CanViewBreeds = 'CanViewBreeds',
    CreateBreeds = 'CreateBreeds',
    BrowseBreeds = 'BrowseBreeds',
    EditBreeds = 'EditBreeds',
    DeleteBreeds = 'DeleteBreeds',
  
    // Adoption Applications
    CanViewAdoptionApplications = 'CanViewAdoptionApplications',
    CreateAdoptionApplications = 'CreateAdoptionApplications',
    BrowseAdoptionApplications = 'BrowseAdoptionApplications',
    EditAdoptionApplications = 'EditAdoptionApplications',
    DeleteAdoptionApplications = 'DeleteAdoptionApplications',
  
    // Files
    CanViewFiles = 'CanViewFiles',
    CreateFiles = 'CreateFiles',
    BrowseFiles = 'BrowseFiles',
    EditFiles = 'EditFiles',
    DeleteFiles = 'DeleteFiles',
  
    // Conversations
    CanViewConversations = 'CanViewConversations',
    CreateConversations = 'CreateConversations',
    BrowseConversations = 'BrowseConversations',
    EditConversations = 'EditConversations',
    DeleteConversations = 'DeleteConversations',
  
    // Messages
    CanViewMessages = 'CanViewMessages',
    CreateMessages = 'CreateMessages',
    BrowseMessages = 'BrowseMessages',
    EditMessages = 'EditMessages',
    DeleteMessages = 'DeleteMessages',
  
    // Notifications
    CanViewNotifications = 'CanViewNotifications',
    CreateNotifications = 'CreateNotifications',
    BrowseNotifications = 'BrowseNotifications',
    EditNotifications = 'EditNotifications',
    DeleteNotifications = 'DeleteNotifications',
  
    // Reports
    CanViewReports = 'CanViewReports',
    CreateReports = 'CreateReports',
    BrowseReports = 'BrowseReports',
    EditReports = 'EditReports',
    DeleteReports = 'DeleteReports'
  }